+++ snooplyze/notifier/__init__.py
@@ -1,6 +1,7 @@
-from .notifier import Notifier
+from .notifier import Notifier, NotifierType
 from .console_notifier import ConsoleNotifier
-from .file_notifier import FileNotifier
+from .flatfile_notifier import FlatFileNotifier
+from .email_notifier import EmailNotifier
 
-__all__ = ["Notifier", 
-           "ConsoleNotifier", "FileNotifier"]
+__all__ = ["Notifier", "NotifierType",
+           "ConsoleNotifier", "FlatFileNotifier", "EmailNotifier"]