+++ snooplyze/notifier/console_notifier.py
@@ -1,4 +1,10 @@
-class ConsoleNotifier:
+from notifier import Notifier
+
+class ConsoleNotifier(Notifier):
+
+    def __init__(self, page_name: str, content_time: str):
+        self.page_name = page_name
+        self.content_time = content_time
     
     def notify(self, message):
-        print(f"🔔 Notification: {message}")
+        print(f"🔔 {self.content_time} | {self.page_name} | {message}")